# openapi/bundled.yaml
openapi: 3.1.0
info:
  title: Chat API
  version: 0.1.0
  description: Slack-like communication application API
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/user-groups:
    post:
      operationId: createUserGroup
      summary: Create a new user group
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserGroupRequest'
      responses:
        '201':
          description: User group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserGroup'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      operationId: listUserGroups
      summary: List user groups
      security:
        - bearerAuth: []
      parameters:
        - name: workspace_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User groups retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  userGroups:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserGroup'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/user-groups/{id}:
    get:
      operationId: getUserGroup
      summary: Get a user group
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User group retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserGroup'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      operationId: updateUserGroup
      summary: Update a user group
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserGroupRequest'
      responses:
        '200':
          description: User group updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserGroup'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      operationId: deleteUserGroup
      summary: Delete a user group
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User group deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/user-groups/{id}/members:
    post:
      operationId: addUserGroupMember
      summary: Add a member to a user group
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMemberRequest'
      responses:
        '201':
          description: Member added
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      operationId: listUserGroupMembers
      summary: List members of a user group
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Members retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  members:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserGroupMember'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      operationId: removeUserGroupMember
      summary: Remove a member from a user group
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: user_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Member removed
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User group or member not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/workspaces/{id}/members:
    get:
      operationId: listMembers
      summary: List workspace members
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of members
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMembersResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      operationId: addMemberByEmail
      summary: Add member to workspace by email
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMemberRequest'
      responses:
        '201':
          description: Member added
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required:
                  - message
  /api/workspaces/{id}/members/{userId}:
    patch:
      operationId: updateMemberRole
      summary: Update member role
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMemberRoleRequest'
      responses:
        '200':
          description: Member role updated
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Member not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      operationId: removeMember
      summary: Remove member from workspace
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Member removed
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Member not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/workspaces/{workspaceId}/search:
    get:
      operationId: searchWorkspace
      summary: Search workspace content
      security:
        - bearerAuth: []
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: filter
          in: query
          schema:
            type: string
            enum:
              - all
              - messages
              - channels
              - users
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: perPage
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
      responses:
        '200':
          description: Search results returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceSearchResponse'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AddMemberRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          enum:
            - admin
            - member
          default: member
      required:
        - email
...
    Attachment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        messageId:
          type: string
          format: uuid
        fileName:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - messageId
        - fileName
        - mimeType
        - sizeBytes
        - createdAt
...
    Channel:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workspaceId:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        isPrivate:
          type: boolean
        channelType:
          type: string
          enum:
            - public
            - private
            - dm
            - group_dm
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        unreadCount:
          type: integer
          description: 未読メッセージ数
          minimum: 0
        hasMention:
          type: boolean
          description: 未読メンションの有無
      required:
        - id
        - workspaceId
        - name
        - isPrivate
        - createdBy
        - createdAt
...
    CreateUserGroupRequest:
      type: object
      properties:
        workspaceId:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
        description:
          type: string
      required:
        - workspaceId
        - name
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    ListMembersResponse:
      type: object
      properties:
        members:
          type: array
          items:
            $ref: '#/components/schemas/MemberInfo'
      required:
        - members
...
    MemberInfo:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
          enum:
            - owner
            - admin
            - member
        joinedAt:
          type: string
          format: date-time
      required:
        - userId
        - email
        - displayName
        - role
        - joinedAt
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        channelId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        parentId:
          type: string
          format: uuid
          nullable: true
        body:
          type: string
        createdAt:
          type: string
          format: date-time
        editedAt:
          type: string
          format: date-time
          nullable: true
        deletedAt:
          type: string
          format: date-time
          nullable: true
        isDeleted:
          type: boolean
        deletedBy:
          type: object
          nullable: true
          properties:
            id:
              type: string
              format: uuid
            displayName:
              type: string
            avatarUrl:
              type: string
              nullable: true
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
      required:
        - id
        - channelId
        - userId
        - body
        - createdAt
        - isDeleted
...
    PaginatedChannels:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Channel'
        total:
          type: integer
          minimum: 0
        page:
          type: integer
          minimum: 1
        perPage:
          type: integer
          minimum: 1
        hasMore:
          type: boolean
      required:
        - items
        - total
        - page
        - perPage
        - hasMore
    PaginatedMessages:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        total:
          type: integer
          minimum: 0
        page:
          type: integer
          minimum: 1
        perPage:
          type: integer
          minimum: 1
        hasMore:
          type: boolean
      required:
        - items
        - total
        - page
        - perPage
        - hasMore
    PaginatedUsers:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MemberInfo'
        total:
          type: integer
          minimum: 0
        page:
          type: integer
          minimum: 1
        perPage:
          type: integer
          minimum: 1
        hasMore:
          type: boolean
      required:
        - items
        - total
        - page
        - perPage
        - hasMore
...
    UpdateMemberRoleRequest:
      type: object
      properties:
        role:
          type: string
          enum:
            - admin
            - member
      required:
        - role
...
    UpdateUserGroupRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
      required:
        - name
...
    UserGroup:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workspaceId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - workspaceId
        - name
        - createdBy
        - createdAt
        - updatedAt
    UserGroupMember:
      type: object
      properties:
        groupId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        joinedAt:
          type: string
          format: date-time
      required:
        - groupId
        - userId
        - joinedAt
...
    WorkspaceSearchResponse:
      type: object
      properties:
        messages:
          $ref: '#/components/schemas/PaginatedMessages'
        channels:
          $ref: '#/components/schemas/PaginatedChannels'
        users:
          $ref: '#/components/schemas/PaginatedUsers'
      required:
        - messages
        - channels
        - users